+++ PS7-x64/Modules/AdminToolbox/Public/Open-AdminMMC.ps1
@@ -1,5 +1,4 @@
 Function Open-AdminMMC {
-
     <#
     .DESCRIPTION
     Opens an mmc concsole with most commonly used MMC Snap-ins for admin functions