+++ PS7-x64/Modules/AdminToolbox.FileManagement/ChangeLog.md
@@ -13,3 +13,4 @@
 * **1.2.0.0** Tested and marked Core compatbile
 * **1.2.1.0** Modified Remove-DisabledADProfiles to use a Try/Catch
 * **1.3.0.0** Added Get-ShareReport Function that combines ntfs and share permissions reports
+* **1.4.0.0** Added Functions get-filename and get-foldername