+++ PS7-x64/Modules/AdminToolbox.Office365/ChangeLog.md
@@ -21,3 +21,4 @@
 * **1.2.0.0** Added multiple functions and updated get-office365 function
 * **1.3.0.0** Added multiple functions and updated get-office365 function
 * **1.3.0.1** Fixed errors in the get-office365 function
+* **2.0.0.0** Updated Get-Office365 and modified the logical structure of the module