+++ PS7-x64/Modules/AdminToolbox/ChangeLog.md
@@ -128,3 +128,4 @@
 * **7.1.2.0**  Updated Required module versions
 * **7.1.3.0**  Updated multiple module dependencies and updated get-info function
 * **7.1.4.0**  Updated multiple module dependencies
+* **8.0.0.0**  Moved two functions to the file management module. Added cmder to the module.